refactor: 기존 MonthlyReport 관련 명칭을 Disc로 일괄 변경
