+++ src/main/java/com/coredisc/domain/disc/DiscRepository.java
@@ -1,8 +1,8 @@
-package com.coredisc.domain.monthlyReport;
+package com.coredisc.domain.disc;
 
 import com.coredisc.domain.member.Member;
 
-public interface MonthlyReportRepository {
+public interface DiscRepository {
 
     Long countByMember(Member member);
 }
