+++ src/main/java/com/coredisc/domain/member/Member.java
@@ -8,7 +8,7 @@ import com.coredisc.domain.common.enums.OauthType;
 import com.coredisc.domain.common.enums.Role;
 import com.coredisc.domain.follow.Follow;
 import com.coredisc.domain.mapping.MemberTerms;
-import com.coredisc.domain.monthlyReport.MonthlyReport;
+import com.coredisc.domain.disc.Disc;
 import com.coredisc.domain.profileImg.ProfileImg;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import jakarta.persistence.*;
@@ -83,7 +83,7 @@ public class Member extends BaseEntity {
     private List<MemberTerms> memberTermsList = new ArrayList<>();
 
     @OneToMany(mappedBy = "member", cascade = CascadeType.ALL, orphanRemoval = true)
-    private List<MonthlyReport> monthlyReportList = new ArrayList<>();
+    private List<Disc> discList = new ArrayList<>();
 
     // 이 사용자가 팔로우하는 다른 사용자들과의 관계 목록
     @OneToMany(mappedBy = "follower", cascade = CascadeType.ALL)
