--- src/main/java/com/coredisc/domain/monthlyReport/MonthlyReport.java
@@ -1,26 +1,0 @@
-package com.coredisc.domain.monthlyReport;
-
-import com.coredisc.domain.common.BaseEntity;
-import com.coredisc.domain.member.Member;
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class MonthlyReport extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    // TODO: 다른 필드 추가
-
-    // 연관 관계
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id", nullable = false)
-    private Member member;
-
-}
