+++ src/main/java/com/coredisc/infrastructure/repository/disc/JpaDiscRepository.java
@@ -1,0 +1,10 @@
+package com.coredisc.infrastructure.repository.disc;
+
+import com.coredisc.domain.member.Member;
+import com.coredisc.domain.disc.Disc;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface JpaDiscRepository extends JpaRepository<Disc, Long> {
+
+    Long countByMember(Member member);
+}
